requirements.txt
###### Requirements without Version Specifiers ######
numpy
pandas
sqlalchemy
mysql-connector-python
pymysql
